# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "2.0.0",
        "description": "APIs for PostAway 2 application",
        "title": "PostAway 2 API"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "paths": {
...
        "/api/posts/": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create A Post",
                "description": "Creates a new post with a caption and an image.",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "description": "Caption for the post"
                                    },
                                    "imageURL": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Image file for the post"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Newly Created Post Object"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    },
                    "500": {
                        "description": "Failed To Create Post"
                    }
                }
            },
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "User Posts",
                "description": "Returns all posts of the logged in user.",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Posts"
                    },
                    "404": {
                        "description": "Posts Not Found"
                    },
                    "500": {
                        "description": "Failed To Fetch Posts"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                }
            }
        },
...
    }
}